move hooks and contexts to hooks module
