+++ packages/hooks/src/index.ts
@@ -2,3 +2,5 @@ export { Canvas } from "./Canvas.js"
 export { useCanvas } from "./useCanvas.js"
 export { useSession } from "./useSession.js"
 export { useRoute } from "./useRoute.js"
+export { useConnect, useSigner, useDisconnect } from "./MultichainConnectContext.js"
+export { MultichainConnect } from "./MultichainConnect.js"
